backend/requirements.txt
# Core dependencies
python-dotenv==1.0.0
openai>=1.35.0
tavily-python>=0.5.0
langchain>=0.3.0
langgraph==0.6.8
langchain-openai>=0.2.0

# Backend framework (Phase 4)
fastapi==0.111.0
uvicorn==0.30.1
websockets==12.0

# Development
pytest==8.2.2
httpx==0.27.0

# Backend framework (Phase 4)
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
python-multipart==0.0.9

# Pydantic for data validation
pydantic==2.7.4
pydantic-settings==2.3.4

# Development
pytest==8.2.2
httpx==0.27.0
pytest-asyncio==0.23.7

# Optional but recommended
redis==5.0.4  # For future cache upgrade
aiofiles==23.2.1  # For async file operations